+++ library/src/scripts/embeddedContent/ImageEmbed.tsx
@@ -7,12 +7,12 @@ import React, { RefObject, useEffect, useRef, useState, useLayoutEffect, useCall
 import { IBaseEmbedProps } from "@library/embeddedContent/embedService";
 import { EmbedContainer } from "@library/embeddedContent/EmbedContainer";
 import { EmbedContent } from "@library/embeddedContent/EmbedContent";
-import { IImageMeta, ImageEmbedMenu } from "@library/embeddedContent/menus/ImageEmbedMenu";
+import { IImageMeta, ImageEmbedMenu } from "@rich-editor/editor/pieces/ImageEmbedMenu";
 import { debuglog } from "util";
 import { IDeviceProps, withDevice } from "@library/layout/DeviceContext";
 import { useFocusWatcher } from "@vanilla/react-utils";
 import classNames from "classnames";
-import { embedMenuClasses } from "@library/embeddedContent/menus/embedMenuStyles";
+import { embedMenuClasses } from "@rich-editor/editor/pieces/embedMenuStyles";
 import { unit } from "@library/styles/styleHelpers";
 import { embedContainerVariables } from "@library/embeddedContent/embedStyles";
