+++ plugins/rich-editor/src/scripts/editor/pieces/embedMenuStyles.ts
